+++ lib/design/widgets/workout_page_widgets/add_exercise_modal/sets_list/sets_cards/working_warmup_set_header_counter_widget.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+class WorkingSetCount extends StatelessWidget {
+  const WorkingSetCount({
+    super.key,
+    required this.setNumber,
+  });
+
+  final int? setNumber;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Text(
+          setNumber == null ? "Warm up Set" : "Working Set: ",
+          style: const TextStyle(
+            color: Color.fromRGBO(255, 255, 255, 0.6),
+            fontSize: 11,
+          ),
+        ),
+        Text(
+          setNumber == null ? "" : "$setNumber",
+          style: const TextStyle(
+            fontWeight: FontWeight.bold,
+            color: Color.fromRGBO(255, 255, 255, 1),
+            fontSize: 11,
+          ),
+        ),
+      ],
+    );
+  }
+}
